+++ internal/datasources/projects.go
@@ -100,7 +100,7 @@ func (d *Projects) Read(ctx context.Context, req datasource.ReadRequest, resp *d
 	url := fmt.Sprintf("%s/v4/organizations/%s/projects", d.HostURL, organizationId)
 	cfg := api.EndpointCfg{Url: url, Method: http.MethodGet, SuccessStatus: http.StatusOK}
 
-	response, err := api.GetPaginated[[]api.GetProjectResponse](ctx, d.Client, d.Token, cfg, api.SortById)
+	response, err := api.GetPaginated[[]api.GetProjectResponse](ctx, d.ClientV1, d.Token, cfg, api.SortById)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Error Reading Capella Projects",